LPS-157280 Moving and renaming class because now it is much more specific for upgrades only
